commit before I delete stuff and/or put things behind feature flags
I was writing code (specifically for handling the clear [screen|line]
escape sequences when I realized that `crossterm` has some nice things
I could use here instead of implementing it myself and facing
(potentially) more bugs.

I was about to delete nearly everything I had written for that function
when I realized the code could probably be used when not writing to a
terminal backend. I've been planning to make a GUI version with a
different feature set and would not have a terminal for crossterm to
render to. So I'm making this commit before I start screwing around with
feature flags.

changelog: ignore