refactor(rendering): Implemented overwriting vs pushing to Line
Most of the underlying parsing/handling logic is completed, if not all.
There are a _few_ things I'm not quite sure about yet but I'm waiting to
see how it plays out after hooking up the rendering and being able to
see results/behavior.

All of the tests are passing, added a few more to verify logic in a few
spots. Cursor movement, color parsing, line overwriting and span management/
splitting all seem to be functioning correctly.

I believe the majority of what is left to do is directly related to the
actual rendering to the terminal. Along with figuring out double bufferring
and how to write it up to enable tui 'popups' (like a menu for session
management). Some things to keep in mind:
  - Still need to track the `ScreenBuffer::view_start` with lines
  - After lines are 'out of view', they should be truncated (trim
    whitespace/empty cells) since they won't be written to anymore and
    truncating will help with their memory size in the VecDeque
  - While hooking up the rendering logic, `serial_actor/*.rs` should be
    re-evaluated and refactored - currently looks like a mess. Would be
    nice to split the task functions up where possible to avoid big
    Future structs on heap.
  - Switch to crossbeam for channels? Not sure how they play w/ Tokio
  - Same with parking_lot and mutexes or maybe RwLock where fitting.
    Been awhile since I looked at that code so I don't remember what it
    is or isn't using.
  - Should spend a day pumping out tests for all of the core
    functionality to get a baseline
  - Hook up `test-sericom` so that testing logic can be a little more
    real-world and longer-running 'sessions' that can be programatically
    run and verified.

Lastly, this won't be relevant for a while yet, but still something to
keep thinking about: how and where can `sericom-core` be split up behind
cargo features to enable usage across different frontends i.e. CLI ||
GUI

changelog: ignore