refactor(rendering): Change `Cell::character` to `u8` for memory opt
Decided to change `Cell::character` from `char` to `u8` because `Cell`s
are only dynamically constructed from the `ByteParser` and all text is
verified to be ascii. This means that it would be no problem to store as
`u8` and simply cast the `u8` to a char when printing/writing if needed.

Switching from `char` to `u8` reduces the size and alignment of `Cell`,
and since there will be _loads_ of `Cell`s in memory, I thought this was
a valid change. When using `char` the size of `Cell` is 8 and alignment
is `4`, when using `u8`, the size is `2` and alignment `1`.