refactor(file,repl): Implemented writing session output to file
Lots of changes all around the codebase, would have been nice to make
smaller commits but undoing-rewriting led me on a wild goose chase all
over to polish a few things up and refactor to fit this new system.

Still mutalating 'sericom/src/main.rs' as I play with the REPL and
notice things, or as needed for interop with sericom_core.

Tracing/debugging/logging is still a WIP, trying to find a flow/system
that I like, but a lot of changes in files from this commit were playing
with the tracing/debugging.

*File-writing:* The file-writing task seems to be where it was before this
refactor. A nice improvement that came along with this refactor is that it
will now write lines that have already been parsed - directly from the
ScreenBuffer. This eliminates just writing whatever the session received
(escape sequence galore).

*Testing:* Wired up the testing harness to simulate a serial connection,
important to note that this only works on Linux and don't plan on
implementing it for Windows because I hate Windows.

See 'sericom-core/src/session/handle.rs' for how to use the harness.