refactor(parsing,rendering): Finished replacing ScreenDriver handling
Finished the implementation of the new version of Parsing/Driving. It is
relatively the same, main difference is the use of `#[repr(u8)]` enums
for matching on incoming bytes/ansii codes. Intent was to make
readability easier and ended up making a few things easier/nicer too.

Have yet to really do much testing, I know that the ORIGIN and cursor
positions all over the codebase are 0-based, and they should really be
1-based so that will need to be fixed. Briefly made a test for the new
ByteParser and it seems to be working fine. Really just need to start
writing tests now to try and find any broken/unexpected behaviors.