test: Testing the handling of CSI/C0/C1 sequences
Converted the Position<TermPos> to be 1-based (1,1) is ORIGIN since that
is what control sequences expect. Tested a majority of the new
implementation and fixed things as a result. Still a few things to test
like scrolling and ensuring the view into the ScreenBuffer's buffer is
up-to-date.

Next thing on the agenda is to wire up the interactive session.

changelog: ignore